+++ db/migrate/20231206000006_create_user_bank_accounts.rb
@@ -7,7 +7,7 @@ class CreateUserBankAccounts < ActiveRecord::Migration[7.0]
       t.integer :agency_number
       t.integer :account_number
       t.boolean :active, null: false, default: true
-      t.decimal :balance, null: false, default: 0.0
+      t.integer :balance, null: false, default: 0.0
 
       t.references :user, null: false, foreign_key: true
       t.references :bank, null: false, foreign_key: true
