+++ db/migrate/20240500000001_create_installments.rb
@@ -4,8 +4,8 @@
 class CreateInstallments < ActiveRecord::Migration[7.1]
   def change
     create_table :installments do |t|
-      t.decimal :starting_price, null: false
-      t.decimal :price, null: false
+      t.integer :starting_price, null: false
+      t.integer :price, null: false
       t.integer :number, null: false
       t.integer :month, null: false
       t.integer :year, null: false
